fixed betty

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stdio.h>
 
 /**
  * listint_len - function that returns no of elements in a linked list.
@@ -8,7 +9,7 @@
 
 size_t listint_len(const listint_t *h)
 {
-        size_t c = 0;
+        unsigned int c = 0;
 
         while (h != NULL)
         {
